move sql to analyse_blocking.py
