+++ splink/blocking.py
@@ -242,84 +242,3 @@ def block_using_rules_sql(linker: Linker):
     sql = "union all".join(sqls)
 
     return sql
-
-
-def count_comparisons_from_blocking_rule_sqls(
-    linker: "Linker", blocking_rule: Union[str, "BlockingRule"]
-):
-    if isinstance(blocking_rule, str):
-        blocking_rule = BlockingRule(blocking_rule, sqlglot_dialect=linker._sql_dialect)
-
-    join_conditions = blocking_rule._join_conditions
-
-    l_cols_sel = []
-    r_cols_sel = []
-    l_cols_gb = []
-    r_cols_gb = []
-    using = []
-    for (
-        i,
-        (l_key, r_key),
-    ) in enumerate(join_conditions):
-        l_cols_sel.append(f"{l_key} as key_{i}")
-        r_cols_sel.append(f"{r_key} as key_{i}")
-        l_cols_gb.append(f"{l_key}")
-        r_cols_gb.append(f"{r_key}")
-        using.append(f"key_{i}")
-
-    l_cols_sel = ", ".join(l_cols_sel)
-    r_cols_sel = ", ".join(r_cols_sel)
-    l_cols_gb = ", ".join(l_cols_gb)
-    r_cols_gb = ", ".join(r_cols_gb)
-    using = ", ".join(using)
-
-    sqls = []
-
-    if not join_conditions:
-        sql = """
-        select count(*) * count(*) as count_of_pairwise_comparisons_generated
-        from __splink__df_concat
-
-        """
-        sqls.append(
-            {"sql": sql, "output_table_name": "__splink__total_of_block_counts"}
-        )
-        return sqls
-
-    sql = f"""
-    select {l_cols_sel}, count(*) as count_l
-    from __splink__df_concat
-    group by {l_cols_gb}
-    """
-
-    sqls.append(
-        {"sql": sql, "output_table_name": "__splink__count_comparisons_from_blocking_l"}
-    )
-
-    sql = f"""
-    select {r_cols_sel}, count(*) as count_r
-    from __splink__df_concat
-    group by {r_cols_gb}
-    """
-
-    sqls.append(
-        {"sql": sql, "output_table_name": "__splink__count_comparisons_from_blocking_r"}
-    )
-
-    sql = f"""
-    select *, count_l, count_r, count_l * count_r as block_count
-    from __splink__count_comparisons_from_blocking_l
-    inner join __splink__count_comparisons_from_blocking_r
-    using ({using})
-    """
-
-    sqls.append({"sql": sql, "output_table_name": "__splink__block_counts"})
-
-    sql = """
-    select sum(block_count) as count_of_pairwise_comparisons_generated
-    from __splink__block_counts
-    """
-
-    sqls.append({"sql": sql, "output_table_name": "__splink__total_of_block_counts"})
-
-    return sqls
